challenge 8 about read a measure in meters and shows the convertion to centimeters and millimeters

+++ problems/4th class/challenge8.py
@@ -1,2 +1,2 @@
-x = int(input('type a measure in meters: '))
+x = float(input('type a measure in meters: '))
 print('CM: {}\nMM: {}\n'.format(x * 100, x * 1000))
